Added type changes to reduce build warnings

+++ SBD Proj 1/fileTape.cpp
@@ -8,7 +8,7 @@ fileTape::fileTape(std::vector<dataBlock> dataFile, unsigned int dataLength): da
 record fileTape::getRecord(unsigned int recordIndex) {
 	if(recordIndex >= this->dataLength) throw std::runtime_error("getRecord error: recordIndex(" + std::to_string(recordIndex) + ") out of range(" + std::to_string(this->dataLength) + ")");
 
-	unsigned int blockIndex = ceil((double)recordIndex / (double)BLOCK_SIZE);
+	unsigned int blockIndex = (unsigned int)ceil((double)recordIndex / (double)BLOCK_SIZE);
 	unsigned int offsetIndex = recordIndex - (blockIndex - 1) * BLOCK_SIZE;
 
 	return this->dataFile[blockIndex].getRecord(offsetIndex);
@@ -16,7 +16,7 @@ record fileTape::getRecord(unsigned int recordIndex) {
 void fileTape::setRecord(unsigned int recordIndex, record newRecord) {
 	if (recordIndex >= this->dataLength) throw std::runtime_error("getRecord error: recordIndex(" + std::to_string(recordIndex) + ") out of range(" + std::to_string(this->dataLength) + ")");
 
-	unsigned int blockIndex = ceil((double)recordIndex / (double)BLOCK_SIZE);
+	unsigned int blockIndex = (unsigned int)ceil((double)recordIndex / (double)BLOCK_SIZE);
 	unsigned int offsetIndex = recordIndex - (blockIndex - 1) * BLOCK_SIZE;
 
 	this->dataFile[blockIndex].setRecord(offsetIndex, newRecord);
